search added in nav bar
