+++ frontend/onestopgo/src/home/index.js
@@ -5,7 +5,6 @@ import 'bootstrap/dist/css/bootstrap.css';
 import CategoriesList from "../categories/categories-list";
 import {useDispatch, useSelector} from "react-redux";
 import {getHomeDataThunk} from "./home-thunk";
-import {useNavigate} from "react-router";
 import Loader from "../components/loader";
 import StoreMap from "../store-map";
 import NavBar from "../nav-bar";
@@ -14,34 +13,15 @@ export const HomeComponent = () => {
     const {homeData, loading} = useSelector(state => state.home)
     const dispatch = useDispatch();
 
-    const [searchText, setSearchText] = useState("");
-
-    function handleChange(e) {
-        setSearchText(e.target.value);
-    }
-
     const mockUserData = () => {
         return {
             imageUrl: 'images/user/190015bd-511a-41f5-ac17-f401aef8df46.jpg'
         }
     }
 
-
     useEffect(() => {
         dispatch(getHomeDataThunk())
     }, []);
-
-    const nav = useNavigate()
-
-    const handleOnSearchSubmit = () => {
-        console.log("Search Submitted");
-        nav("/search", {
-            state:
-                {
-                    searchTerm: searchText
-                }
-        })
-    }
 
     if (!homeData) return null;
 
@@ -51,18 +31,6 @@ export const HomeComponent = () => {
             {loading && <Loader/>}
             {!loading &&
                 <>
-                    <nav>
-                        <div className="nav-wrapper green">
-                            <form onSubmit={handleOnSearchSubmit}>
-                                <div className="input-field">
-                                    <input id="search" type="search" onChange={handleChange} required/>
-                                    <label className="label-icon" htmlFor="search"><i
-                                        className="material-icons">search</i></label>
-                                    <i className="material-icons">close</i>
-                                </div>
-                            </form>
-                        </div>
-                    </nav>
                     <StoreList storeArray={homeData.stores}/>
                     <CategoriesList categoriesArray={homeData.categories}/>
                     <StoreMap storeArray={homeData.stores}/>
